fix(migrations): corriger types colonnes et contraintes FK pour production
- products.stock_quantity/min_stock : INT → DECIMAL(10,4)
  (StockService travaille en float — INT tronquait les fractions)
- Product model : cast integer → float sur stock_quantity/min_stock
- purchase_items.quantity : DECIMAL(10,3) → DECIMAL(10,4) (cohérence stock_movements)
- cash_sessions.opened_by : NOT NULL → nullable (impossible de supprimer un utilisateur)
- sale_items : ajout de timestamps() pour traçabilité
- create_customers_table.down() : dropForeignIdFor → dropConstrainedForeignId (explicite)
- Migration ALTER pour mettre à jour la DB locale existante sans migrate:fresh

+++ Modules/Products/app/Models/Product.php
@@ -37,8 +37,8 @@ class Product extends Model
             'margin'         => 'float',
             'margin_rate'    => 'float',
             'markup_rate'    => 'float',
-            'stock_quantity' => 'integer',
-            'min_stock'      => 'integer',
+            'stock_quantity' => 'float',
+            'min_stock'      => 'float',
             'is_active'      => 'boolean',
             'unit'           => ProductUnit::class,
         ];
